// frontend/src/api/openapi/models/DailyCurrencyRates.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Exchange Rates
 * Exchange Rates API proxy for https://exchangerate-api.com
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DailyCurrencyRates
 */
export interface DailyCurrencyRates {
    /**
     * YYYY-MM-DD
     * @type {string}
     * @memberof DailyCurrencyRates
     */
    day: string;
    /**
     * Currency Code
     * @type {string}
     * @memberof DailyCurrencyRates
     */
    from: string;
    /**
     * Currency Code
     * @type {string}
     * @memberof DailyCurrencyRates
     */
    to: string;
    /**
     * Exchange Rate
     * @type {number}
     * @memberof DailyCurrencyRates
     */
    rate: number;
    /**
     * Other Exchange Rates
     * @type {object}
     * @memberof DailyCurrencyRates
     */
    other: object;
}

/**
 * Check if a given object implements the DailyCurrencyRates interface.
 */
export function instanceOfDailyCurrencyRates(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "day" in value;
    isInstance = isInstance && "from" in value;
    isInstance = isInstance && "to" in value;
    isInstance = isInstance && "rate" in value;
    isInstance = isInstance && "other" in value;

    return isInstance;
}

export function DailyCurrencyRatesFromJSON(json: any): DailyCurrencyRates {
    return DailyCurrencyRatesFromJSONTyped(json, false);
}

export function DailyCurrencyRatesFromJSONTyped(json: any, ignoreDiscriminator: boolean): DailyCurrencyRates {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'day': json['day'],
        'from': json['from'],
        'to': json['to'],
        'rate': json['rate'],
        'other': json['other'],
    };
}

export function DailyCurrencyRatesToJSON(value?: DailyCurrencyRates | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'day': value.day,
        'from': value.from,
        'to': value.to,
        'rate': value.rate,
        'other': value.other,
    };
}